Move URL session uploader out, add multi-delegate support
