+++ ios/Classes/SwiftFlutterUploaderPlugin.swift
@@ -10,16 +10,14 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
     static let KEY_TASK_ID = "task_id"
     static let KEY_STATUS = "status"
     static let KEY_PROGRESS = "progress"
-    static let KEY_MAXIMUM_CONCURRENT_TASK = "FUMaximumConnectionsPerHost"
-    static let KEY_MAXIMUM_CONCURRENT_UPLOAD_OPERATION = "FUMaximumUploadOperation"
-    static let KEY_TIMEOUT_IN_SECOND = "FUTimeoutInSeconds"
-    static let KEY_BACKGROUND_SESSION_IDENTIFIER = "chillisource.flutter_uploader.upload.background"
 
     static let KEY_FIELD_NAME = "fieldname"
     static let KEY_PATH = "path"
     static let STEP_UPDATE = 0
     static let DEFAULT_TIMEOUT = 3600.0
 
+    let urlSessionUploader = URLSessionUploader.shared
+    
     let channel: FlutterMethodChannel
     let progressEventChannel: FlutterEventChannel
     let progressHandler: SimpleStreamHandler
@@ -27,13 +25,7 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
     let resultEventChannel: FlutterEventChannel
     let resultHandler: SimpleStreamHandler
     
-    var session: URLSession
-    let queue: OperationQueue
     let taskQueue: DispatchQueue
-    let timeout: Double
-    var backgroundTransferCompletionHander: Any?
-    var uploadedData = [String: Data]()
-    var runningTaskById = [String: UploadTask]()
 
     public static func register(with registrar: FlutterPluginRegistrar) {
         let channel = FlutterMethodChannel(name: SwiftFlutterUploaderPlugin.channelName, binaryMessenger: registrar.messenger())
@@ -56,14 +48,11 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
         self.resultEventChannel = resultEventChannel
         self.resultHandler = SimpleStreamHandler()
         resultEventChannel.setStreamHandler(resultHandler)
-        
-        self.queue = OperationQueue()
-        self.queue.name = "chillisource.flutter_uploader.queue"
-        self.session = URLSession()
+                
         self.taskQueue = DispatchQueue(label: "chillisource.flutter_uploader.dispatch.queue")
-        self.timeout = SwiftFlutterUploaderPlugin.determineTimeout()
         super.init()
-        self.setupSession()
+        
+        urlSessionUploader.addDelegate(self)
     }
 
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
@@ -83,37 +72,8 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
         }
     }
 
-    
     private func setBackgroundHandler(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
         result(nil)
-    }
-
-    private func setupSession() {
-        let mainBundle = Bundle.main
-        var maxConcurrentTasks: NSNumber
-        if let concurrentTasks = mainBundle.object(forInfoDictionaryKey: SwiftFlutterUploaderPlugin.KEY_MAXIMUM_CONCURRENT_TASK) {
-            maxConcurrentTasks = concurrentTasks as! NSNumber
-        } else {
-            maxConcurrentTasks = NSNumber(integerLiteral: 3)
-        }
-
-        NSLog("MAXIMUM_CONCURRENT_TASKS = \(maxConcurrentTasks)")
-
-        var maxUploadOperation: NSNumber
-        if let operationTask = mainBundle.object(forInfoDictionaryKey: SwiftFlutterUploaderPlugin.KEY_MAXIMUM_CONCURRENT_UPLOAD_OPERATION) {
-            maxUploadOperation = operationTask as! NSNumber
-        } else {
-            maxUploadOperation = NSNumber(integerLiteral: 2)
-        }
-
-        NSLog("MAXIMUM_CONCURRENT_UPLOAD_OPERATION = \(maxUploadOperation)")
-
-        self.queue.maxConcurrentOperationCount = maxUploadOperation.intValue
-
-        let sessionConfiguration = URLSessionConfiguration.background(withIdentifier: SwiftFlutterUploaderPlugin.KEY_BACKGROUND_SESSION_IDENTIFIER)
-        sessionConfiguration.httpMaximumConnectionsPerHost = maxConcurrentTasks.intValue
-        sessionConfiguration.timeoutIntervalForRequest = timeout
-        self.session = URLSession(configuration: sessionConfiguration, delegate: self, delegateQueue: queue)
     }
 
     private func enqueueMethodCall(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
@@ -146,12 +106,10 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
         uploadTaskWithURLWithCompletion(url: url, files: files!, method: method, headers: headers, parameters: data, tag: tag, completion: { (task, error) in
             if (error != nil) {
                 result(error!)
-            } else {
-                let uploadTask = task!
-                let taskId = self.identifierForTask(uploadTask)
-                self.runningTaskById[taskId] = UploadTask(taskId: taskId, status: .enqueue, progress: 0, tag: tag)
-                result(taskId)
-                self.sendUpdateProgressForTaskId(taskId, inStatus: .enqueue, andProgress: 0, andTag: tag)
+            } else if let uploadTask = task {
+                result(self.urlSessionUploader.identifierForTask(uploadTask))
+                
+//                self.sendUpdateProgressForTaskId(taskId, inStatus: .enqueue, andProgress: 0, andTag: tag)
             }
         })
     }
@@ -196,12 +154,8 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
         binaryUploadTaskWithURLWithCompletion(url: url, file: fileUrl, method: method, headers: headers, tag: tag, completion: { (task, error) in
             if (error != nil) {
                 result(error!)
-            } else {
-                let uploadTask = task!
-                let taskId = self.identifierForTask(uploadTask)
-                self.runningTaskById[taskId] = UploadTask(taskId: taskId, status: .enqueue, progress: 0, tag: tag)
-                result(taskId)
-                self.sendUpdateProgressForTaskId(taskId, inStatus: .enqueue, andProgress: 0, andTag: tag)
+            } else if let uploadTask = task {
+                result(self.urlSessionUploader.identifierForTask(uploadTask))
             }
         })
     }
@@ -209,48 +163,16 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
     private func cancelMethodCall(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
         let args = call.arguments as! [String: Any?]
         let taskId = args[SwiftFlutterUploaderPlugin.KEY_TASK_ID] as! String
-        self.cancelWithTaskId(taskId)
+        
+        urlSessionUploader.cancelWithTaskId(taskId)
+
         result(nil)
     }
 
     private func cancelAllMethodCall(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
-        self.cancelAllTasks()
+        urlSessionUploader.cancelAllTasks() 
+
         result(nil)
-    }
-
-    private func cancelWithTaskId(_ taskId: String) {
-        session.getTasksWithCompletionHandler { (_, uploadTasks, _) in
-            uploadTasks.forEach({ (uploadTask) in
-                let state = uploadTask.state
-                let taskIdValue = self.identifierForTask(uploadTask)
-                if taskIdValue == taskId && state == .running {
-                    uploadTask.cancel()
-                    self.sendUpdateProgressForTaskId(taskId, inStatus: .canceled, andProgress: -1, andTag: nil)
-                    return
-                }
-            })
-        }
-    }
-
-    private func cancelAllTasks() {
-        session.getTasksWithCompletionHandler { (_, uploadTasks, _) in
-            uploadTasks.forEach({ (uploadTask) in
-                let state = uploadTask.state
-                let taskId = self.identifierForTask(uploadTask)
-                if state == .running {
-                    uploadTask.cancel()
-                    self.sendUpdateProgressForTaskId(taskId, inStatus: .canceled, andProgress: -1, andTag: nil)
-                }
-            })
-        }
-    }
-
-    private func identifierForTask(_ task: URLSessionUploadTask) -> String {
-        return "\(self.session.configuration.identifier ?? "chillisoure.flutter_uploader").\(task.taskIdentifier)"
-    }
-
-    private func identifierForTask(_ task: URLSessionUploadTask, withSession session: URLSession) -> String {
-        return "\(session.configuration.identifier ?? "chillisoure.flutter_uploader").\(task.taskIdentifier)"
     }
 
     private func binaryUploadTaskWithURLWithCompletion(url: URL,
@@ -269,12 +191,7 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
             }
         }
 
-        request.timeoutInterval = self.timeout
-
-        let uploadTask = self.session.uploadTask(with: request as URLRequest, fromFile: URL(fileURLWithPath: file.path))
-        uploadTask.taskDescription = "\(file.path),KEEP"
-        uploadTask.resume()
-        completionHandler(uploadTask, nil)
+        completionHandler(self.urlSessionUploader.enqueueUploadTask(request as URLRequest, path: file.path), nil)
     }
 
     private func uploadTaskWithURLWithCompletion(url: URL, files: [Any],
@@ -321,8 +238,6 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
                 }
             }
 
-            let tout: Int = Int(self.timeout)
-
             if fileCount <= 0 {
                 completionHandler(nil, flutterError)
             } else {
@@ -348,46 +263,14 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
                     return
                 }
 
-                self.makeRequest(path, url, method, headers, formData.contentType, formData.contentLength, tout, completion: {
+                self.makeRequest(path, url, method, headers, formData.contentType, formData.contentLength, completion: {
                     (task, error) in
                     completionHandler(task, error)
                 })
             }
     }
 
-    private func sendUpdateProgressForTaskId(_ taskId: String, inStatus status: UploadTaskStatus, andProgress progress: Int, andTag tag: String?) {
-        progressHandler.add([
-            SwiftFlutterUploaderPlugin.KEY_TASK_ID: taskId,
-            SwiftFlutterUploaderPlugin.KEY_STATUS: NSNumber(integerLiteral: status.rawValue),
-            SwiftFlutterUploaderPlugin.KEY_PROGRESS: NSNumber(integerLiteral: progress),
-            "tag": (tag ?? NSNull()) as Any
-        ])
-    }
-
-    private func sendUploadFailedForTaskId(_ taskId: String, inStatus status: UploadTaskStatus, statusCode: Int, error: FlutterError, tag: String?) {
-        resultHandler.add([
-            SwiftFlutterUploaderPlugin.KEY_TASK_ID: taskId,
-            SwiftFlutterUploaderPlugin.KEY_STATUS: NSNumber(integerLiteral: status.rawValue),
-            "code": error.code,
-            "message": error.message ?? NSNull(),
-            "details": error.details ?? NSNull(),
-            "statusCode": NSNumber(integerLiteral: statusCode),
-            "tag": tag as Any
-        ])
-    }
-
-    private func sendUploadSuccessForTaskId(_ taskId: String, inStatus status: UploadTaskStatus, message: String?, statusCode: Int, headers: [String: Any], tag: String?) {
-        resultHandler.add([
-            SwiftFlutterUploaderPlugin.KEY_TASK_ID: taskId,
-            SwiftFlutterUploaderPlugin.KEY_STATUS: NSNumber(integerLiteral: status.rawValue),
-            "message": message ?? NSNull(),
-            "statusCode": statusCode,
-            "headers": headers,
-            "tag": tag ?? NSNull()
-        ])
-    }
-
-    private func makeRequest(_ path: String, _ url: URL, _ method: String, _ headers: [String: Any?]?, _ contentType: String, _ contentLength: UInt64, _ timeout: Int, completion completionHandler: (URLSessionUploadTask?, FlutterError?) -> Void) {
+    private func makeRequest(_ path: String, _ url: URL, _ method: String, _ headers: [String: Any?]?, _ contentType: String, _ contentLength: UInt64, completion completionHandler: (URLSessionUploadTask?, FlutterError?) -> Void) {
 
         let request = NSMutableURLRequest(url: url)
         request.httpMethod = method
@@ -403,8 +286,6 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
             }
         }
 
-        request.timeoutInterval = Double(timeout)
-
         let fm = FileManager.default
 
         if !fm.fileExists(atPath: path) {
@@ -412,118 +293,7 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
             return
         }
 
-        let uploadTask = self.session.uploadTask(with: request as URLRequest, fromFile: URL(fileURLWithPath: path))
-        uploadTask.taskDescription = "\(path),DELETE"
-        uploadTask.resume()
-        completionHandler(uploadTask, nil)
-    }
-
-    public func urlSession(_ session: URLSession, task: URLSessionTask, didCompleteWithError error: Error?) {
-
-        guard let uploadTask = task as? URLSessionUploadTask else {
-            NSLog("URLSessionDidCompleteWithError: not an uplaod task")
-            return
-        }
-
-        let taskId = identifierForTask(uploadTask, withSession: session)
-        let runningTask = self.runningTaskById[taskId]
-        let tag = runningTask?.tag
-
-        if error != nil {
-            NSLog("URLSessionDidCompleteWithError: \(taskId) failed with \(error!.localizedDescription)")
-            var uploadStatus: UploadTaskStatus = .failed
-            switch error! {
-            case URLError.cancelled:
-                uploadStatus = .canceled
-            default:
-                uploadStatus = .failed
-            }
-
-            self.sendUploadFailedForTaskId(taskId, inStatus: uploadStatus, statusCode: 500, error: FlutterError(code: "upload_error", message: error?.localizedDescription, details: Thread.callStackSymbols), tag: tag)
-            self.runningTaskById.removeValue(forKey: taskId)
-            self.uploadedData.removeValue(forKey: taskId)
-            return
-        }
-
-        var hasResponseError = false
-        var response: HTTPURLResponse?
-        var statusCode = 500
-
-        if task.response is HTTPURLResponse {
-            response = task.response as? HTTPURLResponse
-
-            if response != nil {
-                NSLog("URLSessionDidCompleteWithError: \(taskId) with response: \(response!) and status: \(response!.statusCode)")
-                statusCode = response!.statusCode
-                hasResponseError = !isRequestSuccessful(response!.statusCode)
-            }
-        }
-
-        NSLog("URLSessionDidCompleteWithError: upload completed")
-
-        if let pathDescription = uploadTask.taskDescription {
-            let split = pathDescription.split(separator: ",")
-
-            if split.count == 2 {
-                let path = String(split[0])
-                if split[1] == "DELETE" && FileManager.default.fileExists(atPath: path) {
-                    do {
-                        try FileManager.default.removeItem(atPath: path)
-                    } catch {
-                        NSLog("URLSessionDidCompleteWithError: Failed to delete file in path \(path)")
-                    }
-                }
-            }
-        }
-
-        let headers = response?.allHeaderFields
-        var responseHeaders = [String: Any]()
-        if headers != nil {
-            headers!.forEach { (key, value) in
-                if let k = key as? String {
-                    responseHeaders[k] = value
-                }
-            }
-        }
-
-        var data: Data = Data()
-
-        if uploadedData.contains(where: { (key, _) in
-            return key == taskId
-        }) {
-            let d = uploadedData[taskId]
-            if d != nil {
-                data = d!
-            }
-        }
-
-        self.uploadedData.removeValue(forKey: taskId)
-        self.runningTaskById.removeValue(forKey: taskId)
-
-        let dataString = String(data: data, encoding: String.Encoding.utf8)
-        let message = dataString == nil ? "" : dataString!
-        if error == nil && !hasResponseError {
-            NSLog("URLSessionDidCompleteWithError: response: \(message), task: \(uploadTask.state.statusText())")
-            self.sendUploadSuccessForTaskId(taskId, inStatus: .completed, message: message, statusCode: response?.statusCode ?? 200, headers: responseHeaders, tag: tag)
-        } else if hasResponseError {
-            NSLog("URLSessionDidCompleteWithError: task: \(uploadTask.state.statusText()) statusCode: \(response?.statusCode ?? -1), error:\(message), response:\(String(describing: response))")
-            self.sendUploadFailedForTaskId(taskId, inStatus: .failed, statusCode: statusCode, error: FlutterError(code: "upload_error", message: message, details: Thread.callStackSymbols), tag: tag)
-        } else {
-            NSLog("URLSessionDidCompleteWithError: task: \(uploadTask.state.statusText()) statusCode: \(response?.statusCode ?? -1), error:\(error?.localizedDescription ?? "none")")
-            self.sendUploadFailedForTaskId(taskId, inStatus: .failed, statusCode: statusCode, error: FlutterError(code: "upload_error", message: error?.localizedDescription, details: Thread.callStackSymbols), tag: tag)
-        }
-    }
-
-    private func isRequestSuccessful(_ statusCode: Int) -> Bool {
-        return statusCode >= 200 && statusCode <= 299
-    }
-
-    private static func determineTimeout() -> Double {
-        if let timeoutSetting = Bundle.main.object(forInfoDictionaryKey: KEY_TIMEOUT_IN_SECOND) {
-            return (timeoutSetting as! NSNumber).doubleValue
-        } else {
-            return SwiftFlutterUploaderPlugin.DEFAULT_TIMEOUT
-        }
+        completionHandler(urlSessionUploader.enqueueUploadTask(request as URLRequest, path: path), nil)
     }
 }
 
@@ -531,104 +301,52 @@ public class SwiftFlutterUploaderPlugin: NSObject, FlutterPlugin {
 extension SwiftFlutterUploaderPlugin {
     public func application(_ application: UIApplication, handleEventsForBackgroundURLSession identifier: String, completionHandler: @escaping () -> Void) -> Bool {
         NSLog("ApplicationHandleEventsForBackgroundURLSession: \(identifier)")
-        if identifier == SwiftFlutterUploaderPlugin.KEY_BACKGROUND_SESSION_IDENTIFIER {
-            self.backgroundTransferCompletionHander = completionHandler
+        if identifier == URLSessionUploader.KEY_BACKGROUND_SESSION_IDENTIFIER {
+            urlSessionUploader.backgroundTransferCompletionHander = completionHandler
         }
+        
         return true
     }
 
     public func applicationWillTerminate(_ application: UIApplication) {
-        runningTaskById.removeAll()
-        uploadedData.removeAll()
-        queue.cancelAllOperations()
+//        runningTaskById.removeAll()
+//        uploadedData.removeAll()
+//        queue.cancelAllOperations()
     }
 }
 
-extension SwiftFlutterUploaderPlugin : URLSessionTaskDelegate {
-    public func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive data: Data) {
-        NSLog("URLSessionDidReceiveData:")
-
-        guard let uploadTask = dataTask as? URLSessionUploadTask else {
-            NSLog("URLSessionDidReceiveData: not an uplaod task")
-            return
-        }
-        let taskId = identifierForTask(uploadTask, withSession: session)
-
-        if uploadedData.contains(where: { (key, _) in
-            return key == taskId
-        }) {
-            NSLog("URLSessionDidReceiveData: existing data with \(taskId)")
-            if data.count > 0 {
-                self.uploadedData[taskId]?.append(data)
-            }
-        } else {
-            var udata = Data()
-            if(data.count > 0) {
-                udata.append(data)
-            }
-
-            self.uploadedData[taskId] = udata
-        }
+extension SwiftFlutterUploaderPlugin : UploaderDelegate {
+    func uploadProgressed(taskId: String, inStatus: UploadTaskStatus, progress: Int) {
+        progressHandler.add([
+            SwiftFlutterUploaderPlugin.KEY_TASK_ID: taskId,
+            SwiftFlutterUploaderPlugin.KEY_STATUS: inStatus.rawValue,
+            SwiftFlutterUploaderPlugin.KEY_PROGRESS: progress,
+//            "tag": (tag ?? NSNull()) as Any
+        ])
     }
+    
+    func uploadCompleted(taskId: String, message: String, statusCode: Int, headers: [String : Any]) {
+        resultHandler.add([
+            SwiftFlutterUploaderPlugin.KEY_TASK_ID: taskId,
+            SwiftFlutterUploaderPlugin.KEY_STATUS: UploadTaskStatus.completed.rawValue,
+            "message": message,
+            "statusCode": statusCode,
+            "headers": headers,
+//            "tag": tag ?? NSNull()
+        ])
 
-    public func urlSession(_ session: URLSession, didBecomeInvalidWithError error: Error?) {
-        NSLog("URLSessionDidBecomeInvalidWithError:")
     }
-
-    public func urlSession(_ session: URLSession, taskIsWaitingForConnectivity task: URLSessionTask) {
-        NSLog("URLSessionTaskIsWaitingForConnectivity:")
+    
+    func uploadFailed(taskId: String, inStatus: UploadTaskStatus, statusCode: Int, errorCode: String, errorMessage: String, errorStackTrace: [String]) {
+        resultHandler.add([
+            SwiftFlutterUploaderPlugin.KEY_TASK_ID: taskId,
+            SwiftFlutterUploaderPlugin.KEY_STATUS: inStatus.rawValue,
+            "statusCode": NSNumber(integerLiteral: statusCode),
+            "code": errorCode,
+            "message": errorMessage,
+            "details": errorStackTrace
+        ])
     }
-
-    public func urlSession(_ session: URLSession, task: URLSessionTask, didSendBodyData bytesSent: Int64, totalBytesSent: Int64, totalBytesExpectedToSend: Int64) {
-
-        if totalBytesExpectedToSend == NSURLSessionTransferSizeUnknown {
-            NSLog("Unknown transfer size")
-        } else {
-            guard let uploadTask = task as? URLSessionUploadTask else {
-                NSLog("URLSessionDidSendBodyData: an not uplaod task")
-                return
-            }
-
-            let taskId = identifierForTask(uploadTask, withSession: session)
-            let bytesExpectedToSend = Double(integerLiteral: totalBytesExpectedToSend)
-            let tBytesSent = Double(integerLiteral: totalBytesSent)
-            let progress = round(Double(tBytesSent / bytesExpectedToSend * 100))
-            let runningTask = self.runningTaskById[taskId]
-            NSLog("URLSessionDidSendBodyData: taskId: \(taskId), byteSent: \(bytesSent), totalBytesSent: \(totalBytesSent), totalBytesExpectedToSend: \(totalBytesExpectedToSend), progress:\(progress)")
-
-            if runningTask != nil {
-                let isRunning: (Int, Int, Int) -> Bool = {
-                    (current, previous, step) in
-                    let prev = previous + step
-                    return (current == 0 || current > prev || current >= 100) &&  current != previous
-                }
-
-                if isRunning(Int(progress), runningTask!.progress, SwiftFlutterUploaderPlugin.STEP_UPDATE) {
-                    self.sendUpdateProgressForTaskId(taskId, inStatus: .running, andProgress: Int(progress), andTag: runningTask?.tag)
-                    self.runningTaskById[taskId] = UploadTask(taskId: taskId, status: .running, progress: Int(progress), tag: runningTask?.tag)
-                }
-            }
-        }
-    }
-
-    public func urlSessionDidFinishEvents(forBackgroundURLSession session: URLSession) {
-        NSLog("URLSessionDidFinishEvents:")
-        session.getTasksWithCompletionHandler { (_, uploadTasks, _) in
-            if( uploadTasks.count == 0) {
-                NSLog("all upload tasks have been completed")
-
-                if(self.backgroundTransferCompletionHander != nil) {
-                    let completionHandler = self.backgroundTransferCompletionHander as! () -> Void
-
-                    self.backgroundTransferCompletionHander = nil
-
-                    OperationQueue.main.addOperation({
-                        completionHandler()
-
-                        //
-                    })
-                }
-            }
-        }
-    }
+    
+    
 }
